v2.2.2.0
- Massive refactoring and asynchronous fixes
- Disabled file operations from displaying in the install info label, and they now only display in the log
- Disabled selection validation, as previous issues warranting it have since been fixed

+++ CreamInstaller/Classes/SteamCMD.cs
@@ -1,6 +1,4 @@
-﻿using Gameloop.Vdf;
-using Gameloop.Vdf.Linq;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
@@ -10,6 +8,9 @@ using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+
+using Gameloop.Vdf;
+using Gameloop.Vdf.Linq;
 
 namespace CreamInstaller
 {
@@ -26,12 +27,9 @@ namespace CreamInstaller
         internal static readonly Version MinimumAppInfoVersion = Version.Parse("2.0.3.2");
         internal static readonly string AppInfoVersionPath = AppInfoPath + @"\version.txt";
 
-        internal static string Run(string command)
+        internal static async Task<string> Run(string command) => await Task.Run(() =>
         {
-            if (Program.Canceled)
-            {
-                return "";
-            }
+            if (Program.Canceled) return "";
             List<string> logs = new();
             ProcessStartInfo processStartInfo = new()
             {
@@ -55,7 +53,7 @@ namespace CreamInstaller
                 process.WaitForExit();
             }
             return string.Join("\r\n", logs);
-        }
+        });
 
         internal static async Task Setup()
         {
@@ -67,56 +65,34 @@ namespace CreamInstaller
                     byte[] file = await httpClient.GetByteArrayAsync("https://steamcdn-a.akamaihd.net/client/installer/steamcmd.zip");
                     file.Write(ArchivePath);
                 }
-
                 ZipFile.ExtractToDirectory(ArchivePath, DirectoryPath);
                 File.Delete(ArchivePath);
             }
-            if (File.Exists(AppCacheAppInfoPath))
-            {
-                File.Delete(AppCacheAppInfoPath);
-            }
-
+            if (File.Exists(AppCacheAppInfoPath)) File.Delete(AppCacheAppInfoPath);
             if (!File.Exists(AppInfoVersionPath) || !Version.TryParse(File.ReadAllText(AppInfoVersionPath, Encoding.UTF8), out Version version) || version < MinimumAppInfoVersion)
             {
-                if (Directory.Exists(AppInfoPath))
-                {
-                    Directory.Delete(AppInfoPath, true);
-                }
-
+                if (Directory.Exists(AppInfoPath)) Directory.Delete(AppInfoPath, true);
                 Directory.CreateDirectory(AppInfoPath);
                 File.WriteAllText(AppInfoVersionPath, Application.ProductVersion, Encoding.UTF8);
             }
-            if (!File.Exists(DllPath))
-            {
-                Run($@"+quit");
-            }
+            if (!File.Exists(DllPath)) await Run($@"+quit");
         }
 
-        internal static bool GetAppInfo(int appId, out VProperty appInfo, string branch = "public", int buildId = 0)
+        internal static async Task<VProperty> GetAppInfo(int appId, string branch = "public", int buildId = 0)
         {
-            appInfo = null;
-            if (Program.Canceled)
-            {
-                return false;
-            }
-
+            VProperty appInfo = null;
+            if (Program.Canceled) return null;
             string output;
             string appUpdatePath = $@"{AppInfoPath}\{appId}";
             string appUpdateFile = $@"{appUpdatePath}\appinfo.txt";
         restart:
-            if (Program.Canceled)
-            {
-                return false;
-            }
-
+            if (Program.Canceled) return null;
             if (Directory.Exists(appUpdatePath) && File.Exists(appUpdateFile))
-            {
                 output = File.ReadAllText(appUpdateFile, Encoding.UTF8);
-            }
             else
             {
-                Run($@"+@ShutdownOnFailedCommand 0 +login anonymous +app_info_print {appId} +force_install_dir {appUpdatePath} +app_update 4 +quit");
-                output = Run($@"+@ShutdownOnFailedCommand 0 +login anonymous +app_info_print {appId} +quit");
+                await Run($@"+@ShutdownOnFailedCommand 0 +login anonymous +app_info_print {appId} +force_install_dir {appUpdatePath} +app_update 4 +quit");
+                output = await Run($@"+@ShutdownOnFailedCommand 0 +login anonymous +app_info_print {appId} +quit");
                 int openBracket = output.IndexOf("{");
                 int closeBracket = output.LastIndexOf("}");
                 if (openBracket != -1 && closeBracket != -1)
@@ -125,11 +101,7 @@ namespace CreamInstaller
                     File.WriteAllText(appUpdateFile, output, Encoding.UTF8);
                 }
             }
-            if (Program.Canceled || output is null)
-            {
-                return false;
-            }
-
+            if (Program.Canceled || output is null) return null;
             try { appInfo = VdfConvert.Deserialize(output); }
             catch
             {
@@ -139,85 +111,46 @@ namespace CreamInstaller
                     goto restart;
                 }
             }
-            if (appInfo.Value is VValue)
-            {
-                goto restart;
-            }
-
+            if (appInfo.Value is VValue) goto restart;
             if (appInfo is null || (appInfo.Value is not VValue && appInfo.Value.Children().ToList().Count == 0))
-            {
-                return true;
-            }
-
+                return appInfo;
             VToken type = appInfo.Value is VValue ? null : appInfo.Value?["common"]?["type"];
             if (type is null || type.ToString() == "Game")
             {
                 string buildid = appInfo.Value is VValue ? null : appInfo.Value["depots"]?["branches"]?[branch]?["buildid"]?.ToString();
-                if (buildid is null && type is not null)
-                {
-                    return true;
-                }
+                if (buildid is null && type is not null) return appInfo;
                 if (type is null || int.Parse(buildid) < buildId)
                 {
-                    foreach (int id in ParseDlcAppIds(appInfo))
+                    List<int> dlcAppIds = await ParseDlcAppIds(appInfo);
+                    foreach (int id in dlcAppIds)
                     {
                         string dlcAppUpdatePath = $@"{AppInfoPath}\{id}";
-                        if (Directory.Exists(dlcAppUpdatePath))
-                        {
-                            Directory.Delete(dlcAppUpdatePath, true);
-                        }
+                        if (Directory.Exists(dlcAppUpdatePath)) Directory.Delete(dlcAppUpdatePath, true);
                     }
-                    if (Directory.Exists(appUpdatePath))
-                    {
-                        Directory.Delete(appUpdatePath, true);
-                    }
+                    if (Directory.Exists(appUpdatePath)) Directory.Delete(appUpdatePath, true);
                     goto restart;
                 }
             }
-            return true;
+            return appInfo;
         }
 
-        internal static List<int> ParseDlcAppIds(VProperty appInfo)
+        internal static async Task<List<int>> ParseDlcAppIds(VProperty appInfo) => await Task.Run(() =>
         {
             List<int> dlcIds = new();
-            if (appInfo is not VProperty)
-            {
-                return dlcIds;
-            }
-
+            if (appInfo is not VProperty) return dlcIds;
             if (appInfo.Value["extended"] is not null)
-            {
                 foreach (VProperty property in appInfo.Value["extended"])
-                {
                     if (property.Key.ToString() == "listofdlc")
-                    {
                         foreach (string id in property.Value.ToString().Split(","))
-                        {
                             if (!dlcIds.Contains(int.Parse(id)))
-                            {
                                 dlcIds.Add(int.Parse(id));
-                            }
-                        }
-                    }
-                }
-            }
-
             if (appInfo.Value["depots"] is not null)
-            {
                 foreach (VProperty _property in appInfo.Value["depots"])
-                {
                     if (int.TryParse(_property.Key.ToString(), out int _))
-                    {
                         if (int.TryParse(_property.Value?["dlcappid"]?.ToString(), out int appid) && !dlcIds.Contains(appid))
-                        {
                             dlcIds.Add(appid);
-                        }
-                    }
-                }
-            }
-
             return dlcIds;
-        }
+        });
 
         internal static async Task Kill()
         {
@@ -227,10 +160,7 @@ namespace CreamInstaller
                 process.Kill();
                 tasks.Add(Task.Run(() => process.WaitForExit()));
             }
-            foreach (Task task in tasks)
-            {
-                await task;
-            }
+            foreach (Task task in tasks) await task;
         }
 
         internal static void Dispose()
